requirements.txt
appnope==0.1.3
argcomplete==2.0.0
asgiref==3.5.2
asttokens==2.0.8
autopep8==2.0.0
backcall==0.2.0
boto3==1.24.34
botocore==1.27.34
certifi==2022.6.15
cfn-flip==1.3.0
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
durationpy==0.5
executing==1.1.1
gunicorn==20.1.0
hjson==3.1.0
idna==3.3
inflection==0.5.1
ipython==8.5.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
placebo==0.9.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==7.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
traitlets==5.4.0
troposphere==4.1.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.2.2
wsgi-request-logger==0.4.6
zappa==0.55.0
